+++ src/utils/help.ts
@@ -52,11 +52,6 @@ export function printProjectHelp(): void {
   console.log(help);
 }
 
-export function printReplHelp(): void {
-  const help = renderHelpTemplate("repl");
-  console.log(help);
-}
-
 export function printTemplateListHelp(): void {
   const help = renderHelpTemplate("templatelist");
   console.log(help);